Attempted to use quaternions in shader. Reverted back.
My theory was that because quaternions are smaller and their calculations are fairly cheap it would be faster to send them to the shader and do the transformation calculation there, rather than turning them into matrices first. Turns out that despite less data being sent, the extra calculations being done on the GPU makes the rendering loop marginally slower.